Made a spec more robust.
It was testing for key=value&key2=value2, but on my machine the hash
ordering caused it to be output as key2=value2&key=value. Since it is
fairly ambiguous as to the order of key/value pairs you'll get back from
Hash#each, the test is more resilient to that now.