Added a bunch of code which adds a follow_location option to Typhoeus::Request objects. This really just involved adding an accessor to Request objects, and making sure it gets passed through to the Easy objects that actually perform the request.
* Added action to the test server which always returns a redirect. Intended
for testing the follow_location option.

* Implemented couple of specs for existing functionality to verify that
the follow_location option on Typhoeus::Easy does indeed cause Typhoeus
to follow redirects.

* Added follow_location option to Typhoeus::Request.

* Added code to retrieve the follow_location setting on a Typhoeus::Request,
and pass it through to the Typhoeus::Easy object actually performing the
request, plus a simple spec to just verify that the option is being set properly
when set on a Request queued up in Hydra.