Add all 3 .gitignore files to the gemspec.
If someone does "rake gems:unpack" to copy typhoeus into their Rails app
and then "rake gems:build" to compile it, the generated files (.o, .bundle, etc.)
should not be checked in to source control. If they are in source control
and you deploy to a server with a different architecture, the server
will have incorrectly built native files.

Adding the (already existing) .gitignore files to the gemspec prevents
this problem.

Also added *.gem to .gitignore so the local .gem file, if any, doesn't get added
to source control.