Handle more than >1024 open file descriptors.
The change is because when you use more than 1024 file descriptors at a
time, one of them must be numbered >1024. By default, the FD_* macros
used when calling select() just silently corrupt random data if any
single file descriptor is >1024. This is to say, it doesn't matter if
you're only watching one file descriptor, if it is numbered >1024,
behavior is undefined. The fix attempts to remedy this problem for up to
2^19=524288 file descriptors.