Don't stub Ruby core classes without good reason
It is a very poor idea to use String.new as a spy for callback
invocations.  The String.new is a Ruby core method and may be called by
many unrelated methods.  This counter was totally unreliable and
responsible for test failures.

That has been replaced with a true double.  Further tests which check
the value yielded to the block may be useful.