Allow verifyhost to be independent of verifypeer
Prior to this change it wasn't possible to control libcurl's
`ssl_verifyhost` setting independently to its `ssl_verifypeer` setting
when using Typhoeus as a Faraday adapter. It's a not uncommon
situation to want the peer certificate and its chain to be verified, but
to not want to enforce the certicate's CN/SAN-DNS entries match the
hostname of the peer; hence these being two separate options in libcurl.

This change keeps the existing default behaviour of setting
`ssl_verifyhost` to 2 if Faraday's ssl `verify` option is true, and 0
otherwise. However if the `verify` setting is true and `ssl_verifyhost`
has been supplied in the `adapter_options`, then the supplied value is
used.

For example:

```
Faraday.new(url, ssl: {verify: true}) do |faraday|
  faraday.adapter  :typhoeus, ssl_verifyhost: 0
end
```

While it would make sense for the `ssl_verifyhost` option to be supplied
as part of the ssl parameters to the adapter itself, that requires
support in Faraday. I think this approach follows the expected behaviour
of adapters accepting non-standard Faraday options in the
`adapter_options` hash.