Fix case insensitive header handling when response had mixed cases
If the response contained multiple headers of the same name but in mixed
cases (lower case vs proper case, etc), it could break the ability to
access the response headers in a case insensitive fashion.

The basic issue was that the case insensitive handling only happened at
access time (eg, when calling `[]`), but the underlying data being
stored could have mixed cases, which interfered with the ability for
`set_value` to detect when there was a pre-existing header of the same
name that needed to be turned into an array.

So this fixes the issue by having `set_value` operate on the "sanitized"
(aka case insensitive) hash first, and then mirroring that to the `self`
hash afterwards. This maintains compatibility so `self` has copies of
the data using the original key casing (although, I think this is mainly
for inspection purposes, so this could potentially be simplified if this
compatibility isn't really needed).