Increase compatibility of case-insensitive access on response headers
The `Response::Header` class that extends `Hash` allowed for
case-insensitive access via the `[]` method, but not via other
hash-related methods like `fetch`, `key?`, or `dig`.

This extends the case-insensitive access for other standard Hash methods
so other ways of accessing the keys also operate in a case insensitive
fashion. This is largely based on methods Rails overrides in their
`ActiveSupport::HashWithIndifferentAccess` class for similar
purposes (although that's for key versus symbol access, not
case-sensitivity, but the principles are the same). This doesn't go
quite as far as `HashWithIndifferentAccess`, since I've just focused on
read-related methods, though, since I don't think write-related methods
on `Response::Header` would really be expected.

This includes case-insensitive support for the following additional
methods:

- `include?` (and standard Ruby aliases `has_key?`, `key?`, `member?`)
- `fetch`
- `dig`
- `assoc`
- `values_at`
- `fetch_values`